Don't consider all 4xx responses as fatal errors
The way HTTP errors are handled means that status codes such as 404s
would result in a generic "something went wrong" kind of error. The
7digital API is going to start returning at least some of these status
codes properly, alongside the existing error, rather than returning a 200.

If we receive, for example, a 404, continue to behave in the same way as
before.

OAuth errors are still considered fatal errors.

Also add a couple of missing assertions into a few tests.